#188 - move video to extractor directory
