+++ goose/crawler.py
@@ -26,10 +26,11 @@ from copy import deepcopy
 from goose.article import Article
 from goose.utils import URLHelper, RawHelper
 from goose.extractors.content import StandardContentExtractor
+from goose.extractors.videos import VideoExtractor
 from goose.cleaners import StandardDocumentCleaner
 from goose.outputformatters import StandardOutputFormatter
 from goose.images.extractors import UpgradedImageIExtractor
-from goose.videos.extractors import VideoExtractor
+
 from goose.network import HtmlFetcher
